ai_service/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
google-genai
google-generativeai
Pillow
gunicorn
